Implement a 'WavesurferTech' Playback Technology
It would probably be best to re-write this further and remove the actual playback logic (initialization and use of the wavesurfer library) from the plugin. However, I wanted to make this change as backwards compatible as possible, so I wired up both the plugin and tech so that the tech just watches the plugin for events that it needs to re-emit. This is important because VideoJS expects events such as 'timeupdate' to be emitted from the tech instance (not the player or a plugin). This fixes text track support.